requirements.txt
Flask==3.0.2
tensorflow==2.15.0
numpy==1.26.4
pillow==10.2.0
scikit-learn==1.4.1
matplotlib==3.8.3
seaborn==0.13.2